// src/client/matrix/schema/gen_types/m_room_canonical_alias.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomCanonicalAlias {
  alias?: string;
  alt_aliases?: string[];
  [k: string]: unknown;
}