// src/client/matrix/schema/gen_types/m_room_join_rules.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomJoinRules {
  allow?: {
    room_id: string;
    type: "m.room_membership";
    [k: string]: unknown;
  }[];
  join_rule: "public" | "knock" | "invite" | "private" | "restricted";
  [k: string]: unknown;
}