// src/client/matrix/schema/gen_types/m_room_power_levels.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomPowerLevels {
  ban?: number;
  events?: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^.".
     */
    [k: string]: number;
  };
  events_default?: number;
  invite?: number;
  kick?: number;
  notifications?: {
    room?: number;
    [k: string]: unknown;
  };
  redact?: number;
  state_default?: number;
  users?: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^.".
     */
    [k: string]: number;
  };
  users_default?: number;
  [k: string]: unknown;
}