// src/client/matrix/schema/gen_types/m_room_avatar.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomAvatar {
  info?: {
    w?: number;
    h?: number;
    mimetype?: string;
    size?: number;
    thumbnail_url?: string;
    thumbnail_info?: {
      w?: number;
      h?: number;
      mimetype?: string;
      size?: number;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  url?: string;
  [k: string]: unknown;
}