// src/client/matrix/schema/gen_types/m_room_pinned_events.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MRoomPinnedEvents {
  pinned: string[];
  [k: string]: unknown;
}